All API tested

+++ src/main/java/com/decimal/rbac/model/dtos/Pagination.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class Pagination {
     Integer totalPages;
-    Long currentPage;
+    Integer currentPage;
     Long elements;
 
 }
